chore: rename `vincentNetworkContext.ts` file to `NetworkContextManager.ts` and `createVincentContracts.ts` to `ContractDataManager.ts`
