+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/rawContractApis/VincentAppFacet/enableAppVersion.spec.ts
@@ -1,4 +1,4 @@
-import { vincentNetworkContext } from '../../../vincentNetworkContext';
+import { vincentNetworkContext } from '../../../NetworkContextManager';
 import { getTestContext } from '../testContext';
 import { enableAppVersion } from './enableAppVersion';
 import { registerNextAppVersion } from './registerNextAppVersion';
