+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/rawContractApis/VincentUserViewFacet/getPermittedAppVersionForPkp.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import { logger } from '../../../../../../shared/logger';
-import { VincentNetworkContext } from '../../../vincentNetworkContext';
-import { createVincentContracts } from '../../utils/createVincentContracts';
+import { VincentNetworkContext } from '../../../NetworkContextManager';
+import { createVincentContracts } from '../../../ContractDataManager';
 
 // Define raw types from the contract
 type RawContractMethod = ReturnType<
